Do not publish the release with 'todo' as description
Before, our Docker images build was dependent on having the tarball
on GitHub, which means publishing the release so unauthenticated users
can download them.

Since a long time, the tarball is downloaded from download.ledgersmb.org;
which means we can now publish the release on Github as soon as it's
complete. The advantage here is that the release then will be first
published with the full description. Currently, the 'todo' description
appears in my timeline, even after the release announcement 'patches it up'.
The hope is that with this commit, the real description appears in people's
timelines instead.